Fully Functional For up to 4 DCMotors

+++ motorhat_block.py
@@ -6,10 +6,10 @@ from Adafruit_MotorHAT import Adafruit_MotorHAT, Adafruit_DCMotor
 class Motorhat(Block):
 
     version = VersionProperty('0.1.0')
-    motor1_speed = IntProperty(title='DC Motor 1 Speed', default=0)
-    motor2_speed = IntProperty(title='DC Motor 2 Speed', default=0)
-    motor3_speed = IntProperty(title='DC Motor 3 Speed', default=0)
-    motor4_speed = IntProperty(title='DC Motor 4 Speed', default=0)
+    motor1_speed = FloatProperty(title='DC Motor 1 Speed', default=0)
+    motor2_speed = FloatProperty(title='DC Motor 2 Speed', default=0)
+    motor3_speed = FloatProperty(title='DC Motor 3 Speed', default=0)
+    motor4_speed = FloatProperty(title='DC Motor 4 Speed', default=0)
 
     def configure(self, context):
         super().configure(context)
@@ -22,7 +22,7 @@ class Motorhat(Block):
                 direction = Adafruit_MotorHAT.FORWARD if speed >= 0 \
                                     else Adafruit_MotorHAT.BACKWARD
                 self.MotorHAT.getMotor(r).run(direction)
-                self.MotorHAT.getMotor(r).setSpeed(abs(speed))
+                self.MotorHAT.getMotor(r).setSpeed(abs(int(speed)))
 
     def stop(self):
         for r in range(1, 5):
